+++ get_version.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+cat VERSION.md